Add new module android-utils. Moved Preferences Manager to android-utils
