+++ accounts/kethereum/build.gradle
@@ -17,7 +17,6 @@ android {
 }
 
 dependencies {
-
     implementation project(":accounts:base")
     implementation project(":app-core")
     api project(":crypto")
