+++ app/build.gradle
@@ -66,6 +66,7 @@ dependencies {
     implementation project(":app-core")
     implementation project(":accounts:base")
     implementation project(":accounts:kethereum")
+    implementation project(":android-utils")
     implementation project(":crypto")
     implementation project(":mnemonic")
