+++ settings.gradle
@@ -3,6 +3,7 @@ include ':accounts:geth'
 include ':accounts:kethereum'
 include ':app'
 include ':app-core'
+include ':android-utils'
 include ':crypto'
 include ':mnemonic'
 include ':utils'
